requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.7.33
botocore==1.10.35
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
docutils==0.14
future==0.16.0
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
pandas==0.23.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
python-telegram-bot==10.1.0
pytz==2018.4
pyzmq==17.0.0
requests==2.18.4
s3transfer==0.1.13
scipy==1.1.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
